Awesome

+++ src/com/pile/block/BlockManager.java
@@ -36,7 +36,7 @@ public class BlockManager {
 		if (xPos < Game.WIDTH && xPos + b.WIDTH > 0) {
 			if (yPos < Game.HEIGHT && yPos + b.HEIGHT > 0) {
 				g.drawImage(b.getImage(), xPos, yPos, null);
-				g.setColor(new Color(0,0,0,51 * b.getDestroyAmount()));
+				g.setColor(new Color(0,0,0, (int) (51 * b.getDestroyAmount())));
 				g.fillRect(xPos, yPos, Block.WIDTH, Block.HEIGHT);
 			}
 		}
